More specific types

+++ src/util.c
@@ -110,7 +110,7 @@ char *read_line(FILE *f, bool downcase, bool trim) {
 		}
 
 		if(downcase == true) {
-			for(int i = 0; i < strlen(line); i++) {
+			for(size_t i = 0; i < strlen(line); i++) {
 				line[i] = tolower(line[i]);
 			}
 		}
